OK: CenteredContent. Loading needs improvements
